Fix IP coupling when zone is renamed or deleted
When a DNS zone is renamed, and an (IPAM) IP address is linked to a "A"
or "AAAA" DNS record from this zone, the dns_name field is updated.

When a DNS zone is deleted, and an (IPAM) IP address is linked to a "A"
or "AAAA" DNS record from this zone, the dns_name field and custom field
"name" are emptied.