Enforce object permissions on coupled DNS record
The coupled record is immediately saved when the "post_clean" signal
is called. All record creation/modification happen in a transaction,
so that object permissions are enforced.

To complete the coupling, the "ipam_ip_address" field must be updated to
point to the IP Address. It can't be done at this this point because
the IP address is not created yet. It's actually done at the end,
when "post_save" is called. At record creation/modification time,
we monkeypatch the "request" object to hold the necessary informations
into a "post_save_ops" attribute.

The semantic of a permission for record modification was not well
defined. Two permission checks are done:
- on the current record,
- on the modified record
The "test_modify_name_with_dns_object_permission" testcase takes
care of this.

NB: 2 tests cases that were duplicated were removed